+++ src/taf/models/mod.rs
@@ -2,12 +2,16 @@
 //!
 //! Contains types and parsing logic implemented for this crate.
 pub mod forecast;
+/// Exposes the `icing` module.
+pub mod icing;
 /// Exposes the `taf` module.
 pub mod taf;
 /// Exposes the `temperature` module.
 pub mod temperature;
 /// Exposes the `time` module.
 pub mod time;
+/// Exposes the `turbulence` module.
+pub mod turbulence;
 /// Exposes the `wind_shear` module.
 pub mod wind_shear;